+++ NEWS.md
@@ -1,3 +1,5 @@
+# fs (development version)
+
 # fs 1.5.1
 
 * Gábor Csárdi is now the maintainer.