+++ NEWS.md
@@ -1,4 +1,4 @@
-# fs (development version)
+# fs 1.5.2
 
 * `file_create()` and `dir_create()` now return the correct path when `...` arguments are used (@davidchall, #333).